Manager (#35)
* Manager 1

* Manager 2

* Manager 3

* Manager 3

* Manager 5

* Manager 6: Remove AutoReentrantLock

* Manager 7: crlf => lf

* Manager 8: Remove puml

* Manager 9

* Manager 10

* Manager 11

* Manager 12

* Manager 13: DeviceIDServiceTest pass

* Manager 14: OrderServiceTest pass

* Manager 15: ControllingSessionServiceTest pass

* Manager 16: ControlledSessionServiceTest pass

+++ ltsvr/src/main/java/cn/lanthing/svr/service/impl/DeviceIDServiceImpl.java
@@ -35,15 +35,10 @@ import cn.lanthing.svr.dao.UnusedIDDao;
 import cn.lanthing.svr.dao.UsedIDDao;
 import cn.lanthing.svr.model.UnusedID;
 import cn.lanthing.svr.model.UsedID;
-import cn.lanthing.svr.model.UsedIDs;
 import cn.lanthing.svr.service.DeviceIDService;
-import cn.lanthing.utils.AutoLock;
-import cn.lanthing.utils.AutoReentrantLock;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import jakarta.annotation.PostConstruct;
 
 import java.util.UUID;
 
@@ -57,33 +52,39 @@ public class DeviceIDServiceImpl implements DeviceIDService {
     @Autowired
     private UsedIDDao usedIDDao;
 
-    private final AutoReentrantLock lock = new AutoReentrantLock();
-
     @Override
     public DeviceCookiePair allocateDeviceID() {
         long deviceID;
         String cookie = UUID.randomUUID().toString();
-        try (AutoLock lockGuard = this.lock.lockAsResource()) {
-            UnusedID entity = unusedIDDao.getNextDeviceID();
-            deviceID = entity.getDeviceID();
+        synchronized (this) {
+            UnusedID unnUsedID = unusedIDDao.getNextDeviceID();
+            if (unnUsedID == null) {
+                log.error("Get next unused deviceID failed");
+                return null;
+            }
+            deviceID = unnUsedID.getDeviceID();
             unusedIDDao.deleteDeviceID(deviceID);
             usedIDDao.addDeviceID(deviceID, cookie);
         }
-        log.info("Allocate device id '{}'", deviceID);
+        log.info("Allocated new deviceID '{}'", deviceID);
         return new DeviceCookiePair(deviceID, cookie);
     }
 
     @Override
-    public UsedID getUsedDeviceID(long deviceID) {
-        try (AutoLock lockGuard = this.lock.lockAsResource()) {
-            return usedIDDao.queryByDeviceID(deviceID);
-        }
+    public synchronized UsedID getUsedDeviceID(long deviceID) {
+        return usedIDDao.queryByDeviceID(deviceID);
     }
 
     @Override
-    public void updateCookie(long deviceID, String cookie) {
-        try (AutoLock lockGuard = this.lock.lockAsResource()) {
-            usedIDDao.updateCookie(deviceID, cookie);
-        }
+    public synchronized void updateCookie(long deviceID, String cookie) {
+        usedIDDao.updateCookie(deviceID, cookie);
     }
+
+    @Override
+    public synchronized DeviceIDStat getDeviceIDStat() {
+        var usedCount = usedIDDao.countID();
+        var unUsedCount = unusedIDDao.countID();
+        return new DeviceIDStat(usedCount, unUsedCount);
+    }
+
 }
